+++ docs/lib/snippets/platforms/new_connect.dart
@@ -1,4 +1,5 @@
 import 'package:drift/drift.dart';
+// ignore: deprecated_member_use
 import 'package:drift/web/worker.dart';
 
 class Approach1 {