+++ drift/test/platforms/web/local_storage_to_indexed_db_test.dart
@@ -5,6 +5,7 @@ library;
 import 'dart:convert';
 
 import 'package:drift/drift.dart';
+// ignore: deprecated_member_use_from_same_package
 import 'package:drift/web.dart';
 import 'package:test/test.dart';