+++ examples/flutter_web_worker_example/lib/src/platform/platform_web.dart
@@ -1,11 +1,18 @@
 import 'package:drift/drift.dart';
-import 'package:drift/web/worker.dart';
+import 'package:drift/wasm.dart';
 import 'package:flutter/foundation.dart';
 
 class PlatformInterface {
   static QueryExecutor createDatabaseConnection(String databaseName) {
-    return DatabaseConnection.delayed(connectToDriftWorker(
-        kReleaseMode ? 'worker.dart.min.js' : 'worker.dart.js',
-        mode: DriftWorkerMode.shared));
+    return DatabaseConnection.delayed(Future(() async {
+      final database = await WasmDatabase.open(
+        databaseName: databaseName,
+        sqlite3Uri: Uri.parse('/sqlite3.wasm'),
+        driftWorkerUri: Uri.parse(
+          kReleaseMode ? '/worker.dart.min.js' : '/worker.dart.js',
+        ),
+      );
+      return database.resolvedExecutor;
+    }));
   }
 }