+++ examples/web_worker_example/web/worker.dart
@@ -1,13 +1,10 @@
-import 'dart:html';
-
-import 'package:drift/web.dart';
-import 'package:drift/web/worker.dart';
+import 'package:drift/wasm.dart';
 
 void main() {
-  WorkerGlobalScope.instance.importScripts('sql-wasm.js');
-
-  driftWorkerMain(() {
-    return WebDatabase.withStorage(DriftWebStorage.indexedDb('worker',
-        migrateFromLocalStorage: false, inWebWorker: true));
+  WasmDatabase.workerMainForOpen(setupAllDatabases: (database) {
+    database.createFunction(
+      functionName: 'my_function',
+      function: (args) => 'Hello from custom drift worker!',
+    );
   });
 }