+++ extras/integration_tests/legacy_web/test/initializer_test.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:drift/drift.dart';
+// ignore: deprecated_member_use
 import 'package:drift/web.dart';
 
 import 'package:test/test.dart';