+++ extras/integration_tests/legacy_web/test/saves_after_migration_regression_test.dart
@@ -2,6 +2,7 @@
 import 'dart:html';
 
 import 'package:drift/drift.dart';
+// ignore: deprecated_member_use
 import 'package:drift/web.dart';
 import 'package:test/test.dart';